requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.2
virtualenv==20.21.0
peewee==3.16.0